Move chat endpoints into separate file
